+++ src/Network/Network/Models/AzureFirewallPolicy/PSAzureFirewallPolicy.cs
@@ -44,6 +44,8 @@ namespace Microsoft.Azure.Commands.Network.Models
 
         public PSAzureFirewallPolicyDnsSettings DnsSettings { get; set; }
 
+        public PSAzureFirewallPolicySqlSetting SqlSetting { get; set; }
+
         public PSAzureFirewallPolicyIntrusionDetection IntrusionDetection { get; set; }
 
         public PSAzureFirewallPolicyTransportSecurity TransportSecurity { get; set; }