+++ src/Network/Network/ChangeLog.md
@@ -19,6 +19,12 @@
 --->
 
 ## Upcoming Release
+* Added new property `SqlSetting` for Azure Firewall Policy cmdlets 
+    - `Get-AzFirewallPolicy` 
+    - `New-AzFirewallPolicy`
+    - `Set-AzFirewallPolicy`
+* Added new to create new `SqlSetting` object for creating Azure Firewall Policy
+    - `New-AzFirewallPolicySqlSetting`
 * Added new cmdlet to support query Load Balancer inbound nat rule port mapping lists for backend addresses
     - `Get-AzLoadBalancerBackendAddressInboundNatRulePortMapping`
     - Also updated cmdlets to support inbound nat rule V2 configurations